requirements.txt
# WhatsApp Fuel Extractor - Python Dependencies
openpyxl>=3.1.2
pandas>=2.0.0
python-dateutil>=2.8.2
schedule>=1.2.0

# CLI with beautiful output
rich>=13.0.0

# Web dashboard
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2

# Uploads to Google Sheets + Database
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
pymysql>=1.1.0
gspread>=6.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
